copied comments from the DOS implementation of Itsy to the Linux one
I made a few minor changes to the formatting of msdos/itsy.asm in
order to make copying the comments over easier.  I changed the names
of registers and the width of data when copying over the comments
(changed AX to EAX, etc), and tried to write equivalent comments where
the code differed by more than the width of the data.  There may still
be typos or cases where the comments don't reflect the actual code.
The resulting binaries for both versions are identical to before any
changes I made, ensuring I didn't accidently change any of the code
itself.